refactor ekp interface
