+++ fdbclient/include/fdbclient/GetEncryptCipherKeys.actor.h
@@ -1,5 +1,5 @@
 /*
- * GetEncryptCipherKeys.actor.cpp
+ * GetEncryptCipherKeys.actor.h
  *
  * This source file is part of the FoundationDB open source project
  *
@@ -17,20 +17,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#pragma once
+#if defined(NO_INTELLISENSE) && !defined(FDBSERVER_GETCIPHERKEYS_ACTOR_G_H)
+#define FDBSERVER_GETCIPHERKEYS_ACTOR_G_H
+#include "fdbserver/GetEncryptCipherKeys.actor.g.h"
+#elif !defined(FDBSERVER_GETCIPHERKEYS_ACTOR_H)
+#define FDBSERVER_GETCIPHERKEYS_ACTOR_H
 
-#include "fdbserver/EncryptKeyProxyInterface.h"
-#include "fdbserver/GetEncryptCipherKeys.h"
+#include "fdbclient/EncryptKeyProxyInterface.h"
+#include "flow/BlobCipher.h"
 #include "flow/IRandom.h"
 
-#include <boost/functional/hash.hpp>
+#include <unordered_map>
+#include <unordered_set>
 
-namespace {
+#include "flow/actorcompiler.h" // This must be the last #include.
 
-Optional<UID> getEncryptKeyProxyId(const Reference<AsyncVar<ServerDBInfo> const>& db) {
-	return db->get().encryptKeyProxy.map<UID>([](EncryptKeyProxyInterface proxy) { return proxy.id(); });
+template <class T>
+Optional<UID> getEncryptKeyProxyId(const Reference<AsyncVar<T> const>& db) {
+	return db->get().encryptKeyProxy.template map<UID>([](EncryptKeyProxyInterface proxy) { return proxy.id(); });
 }
 
-ACTOR Future<Void> onEncryptKeyProxyChange(Reference<AsyncVar<ServerDBInfo> const> db) {
+ACTOR template <class T>
+Future<Void> onEncryptKeyProxyChange(Reference<AsyncVar<T> const> db) {
 	state Optional<UID> previousProxyId = getEncryptKeyProxyId(db);
 	state Optional<UID> currentProxyId;
 	loop {
@@ -46,9 +55,9 @@ ACTOR Future<Void> onEncryptKeyProxyChange(Reference<AsyncVar<ServerDBInfo> cons
 	return Void();
 }
 
-ACTOR Future<EKPGetLatestBaseCipherKeysReply> getUncachedLatestEncryptCipherKeys(
-    Reference<AsyncVar<ServerDBInfo> const> db,
-    EKPGetLatestBaseCipherKeysRequest request) {
+ACTOR template <class T>
+Future<EKPGetLatestBaseCipherKeysReply> getUncachedLatestEncryptCipherKeys(Reference<AsyncVar<T> const> db,
+                                                                           EKPGetLatestBaseCipherKeysRequest request) {
 	Optional<EncryptKeyProxyInterface> proxy = db->get().encryptKeyProxy;
 	if (!proxy.present()) {
 		// Wait for onEncryptKeyProxyChange.
@@ -73,10 +82,12 @@ ACTOR Future<EKPGetLatestBaseCipherKeysReply> getUncachedLatestEncryptCipherKeys
 	}
 }
 
-} // anonymous namespace
-
-ACTOR Future<std::unordered_map<EncryptCipherDomainId, Reference<BlobCipherKey>>> getLatestEncryptCipherKeys(
-    Reference<AsyncVar<ServerDBInfo> const> db,
+// Get latest cipher keys for given encryption domains. It tries to get the cipher keys from local cache.
+// In case of cache miss, it fetches the cipher keys from EncryptKeyProxy and put the result in the local cache
+// before return.
+ACTOR template <class T>
+Future<std::unordered_map<EncryptCipherDomainId, Reference<BlobCipherKey>>> getLatestEncryptCipherKeys(
+    Reference<AsyncVar<T> const> db,
     std::unordered_map<EncryptCipherDomainId, EncryptCipherDomainName> domains) {
 	state Reference<BlobCipherKeyCache> cipherKeyCache = BlobCipherKeyCache::getInstance();
 	state std::unordered_map<EncryptCipherDomainId, Reference<BlobCipherKey>> cipherKeys;
@@ -132,10 +143,9 @@ ACTOR Future<std::unordered_map<EncryptCipherDomainId, Reference<BlobCipherKey>>
 	return cipherKeys;
 }
 
-namespace {
-
-ACTOR Future<EKPGetBaseCipherKeysByIdsReply> getUncachedEncryptCipherKeys(Reference<AsyncVar<ServerDBInfo> const> db,
-                                                                          EKPGetBaseCipherKeysByIdsRequest request) {
+ACTOR template <class T>
+Future<EKPGetBaseCipherKeysByIdsReply> getUncachedEncryptCipherKeys(Reference<AsyncVar<T> const> db,
+                                                                    EKPGetBaseCipherKeysByIdsRequest request) {
 	Optional<EncryptKeyProxyInterface> proxy = db->get().encryptKeyProxy;
 	if (!proxy.present()) {
 		// Wait for onEncryptKeyProxyChange.
@@ -162,10 +172,12 @@ ACTOR Future<EKPGetBaseCipherKeysByIdsReply> getUncachedEncryptCipherKeys(Refere
 
 using BaseCipherIndex = std::pair<EncryptCipherDomainId, EncryptCipherBaseKeyId>;
 
-} // anonymous namespace
-
-ACTOR Future<std::unordered_map<BlobCipherDetails, Reference<BlobCipherKey>>> getEncryptCipherKeys(
-    Reference<AsyncVar<ServerDBInfo> const> db,
+// Get cipher keys specified by the list of cipher details. It tries to get the cipher keys from local cache.
+// In case of cache miss, it fetches the cipher keys from EncryptKeyProxy and put the result in the local cache
+// before return.
+ACTOR template <class T>
+Future<std::unordered_map<BlobCipherDetails, Reference<BlobCipherKey>>> getEncryptCipherKeys(
+    Reference<AsyncVar<T> const> db,
     std::unordered_set<BlobCipherDetails> cipherDetails) {
 	state Reference<BlobCipherKeyCache> cipherKeyCache = BlobCipherKeyCache::getInstance();
 	state std::unordered_map<BlobCipherDetails, Reference<BlobCipherKey>> cipherKeys;
@@ -233,7 +245,15 @@ ACTOR Future<std::unordered_map<BlobCipherDetails, Reference<BlobCipherKey>>> ge
 	return cipherKeys;
 }
 
-ACTOR Future<TextAndHeaderCipherKeys> getLatestSystemEncryptCipherKeys(Reference<AsyncVar<ServerDBInfo> const> db) {
+struct TextAndHeaderCipherKeys {
+	Reference<BlobCipherKey> cipherTextKey;
+	Reference<BlobCipherKey> cipherHeaderKey;
+};
+
+// Helper method to get latest cipher text key and cipher header key for system domain,
+// used for encrypting system data.
+ACTOR template <class T>
+Future<TextAndHeaderCipherKeys> getLatestSystemEncryptCipherKeys(Reference<AsyncVar<T> const> db) {
 	static std::unordered_map<EncryptCipherDomainId, EncryptCipherDomainName> domains = {
 		{ SYSTEM_KEYSPACE_ENCRYPT_DOMAIN_ID, FDB_DEFAULT_ENCRYPT_DOMAIN_NAME },
 		{ ENCRYPT_HEADER_DOMAIN_ID, FDB_DEFAULT_ENCRYPT_DOMAIN_NAME }
@@ -249,8 +269,10 @@ ACTOR Future<TextAndHeaderCipherKeys> getLatestSystemEncryptCipherKeys(Reference
 	return result;
 }
 
-ACTOR Future<TextAndHeaderCipherKeys> getEncryptCipherKeys(Reference<AsyncVar<ServerDBInfo> const> db,
-                                                           BlobCipherEncryptHeader header) {
+// Helper method to get both text cipher key and header cipher key for the given encryption header,
+// used for decrypting given encrypted data with encryption header.
+ACTOR template <class T>
+Future<TextAndHeaderCipherKeys> getEncryptCipherKeys(Reference<AsyncVar<T> const> db, BlobCipherEncryptHeader header) {
 	std::unordered_set<BlobCipherDetails> cipherDetails{ header.cipherTextDetails, header.cipherHeaderDetails };
 	std::unordered_map<BlobCipherDetails, Reference<BlobCipherKey>> cipherKeys =
 	    wait(getEncryptCipherKeys(db, cipherDetails));
@@ -262,3 +284,6 @@ ACTOR Future<TextAndHeaderCipherKeys> getEncryptCipherKeys(Reference<AsyncVar<Se
 	ASSERT(result.cipherHeaderKey.isValid());
 	return result;
 }
+
+#include "flow/unactorcompiler.h"
+#endif
