+++ fdbserver/workloads/EncryptKeyProxyTest.actor.cpp
@@ -19,7 +19,7 @@
  */
 
 #include "fdbrpc/Locality.h"
-#include "fdbserver/EncryptKeyProxyInterface.h"
+#include "fdbclient/EncryptKeyProxyInterface.h"
 #include "fdbserver/Knobs.h"
 #include "fdbserver/ServerDBInfo.actor.h"
 #include "fdbserver/WorkerInterface.actor.h"
